New option for configure script
New command line option "--option" for specifying additional
defines that will be used for the compilation of Circle and
circle-newlib. It is important to have consistent defines
in the compilation of Circle and circle-newlib, otherwise things
like described in https://github.com/probonopd/MiniDexed/issues/39
happen where inconsistent defines in the compilation of the
libraries caused a hang.