Fix creation of ARCH defines
Since Circle commit
https://github.com/rsta2/circle/commit/8f28def1743e8bb87588349137f2fe6508d8b85b
the "ARCH" make macro was no longer self-contained, but depended
on macro "ARCHCPU". This caused the extraction of the compilation flags
from the "ARCH" macro to be malformed, although the newlib build
system apparently tolerated this.

Adapt the configure script to the relevance of the "ARCHCPU" and "ARCH"
macros in Circle's "Rules.mk" file.