Fix problem with parallel builds
A parallel build with make's "-j" command line option resulted in the
warning:

warning: jobserver unavailable: using -j1. Add '+' to parent make rule.

The Circle build then did not run in parallel mode.

Added '+' to the invocation of shell scripts that build the Circle
libraries as advised by the warning. Now the warning is gone and the
Circle build runs in parallel mode if the top-level make is invoked with
"-j".