Minimal SvelteKit app working
Configured the SvelteKit app for single-page application by adding the
"200.html" fallback page.

Modified the 06-socket kernel so that it fulfills the requirements
for a SvelteKit app with static adapter.

Prepared SD card image with build of minimal SvelteKit application
generated by "npx sv create".

Added a script to generate an SD card image with the SvelteKit
application.