Fix socket problems with 06-socket demo
The socket backlog setting in Mongoose was too small for many parallel
requests which caused problems loading the demo application in
Firefox.

Update Circle for a fix in select() behavior.

Update circle-newlib for a modification to recv()'s behavior. The old
behavior caused file descriptor leaks with the demo application.